+++ source/gatery/hlim/coreNodes/Node_Arithmetic.cpp
@@ -226,7 +226,7 @@ void Node_Arithmetic::estimateSignalDelay(SignalDelay &sigDelay)
     }
 }
 
-void Node_Arithmetic::estimateSignalDelayCriticalInput(SignalDelay &sigDelay, unsigned outputPort, unsigned outputBit, unsigned &inputPort, unsigned &inputBit)
+void Node_Arithmetic::estimateSignalDelayCriticalInput(SignalDelay &sigDelay, size_t outputPort, size_t outputBit, size_t &inputPort, size_t &inputBit)
 {
     auto inDelay0 = sigDelay.getDelay(getDriver(0));
     auto inDelay1 = sigDelay.getDelay(getDriver(1));
@@ -234,8 +234,8 @@ void Node_Arithmetic::estimateSignalDelayCriticalInput(SignalDelay &sigDelay, un
     auto width = getOutputConnectionType(0).width;
 
     float maxDelay = 0.0f;
-    unsigned maxIP = 0;
-    unsigned maxIB = 0;
+    size_t maxIP = 0;
+    size_t maxIB = 0;
 
     for (auto i : utils::Range(width))
         if (inDelay0[i] > maxDelay) {
