+++ source/gatery/hlim/coreNodes/Node_Register.h
@@ -67,7 +67,7 @@ namespace gtry::hlim {
 
         virtual void estimateSignalDelay(SignalDelay &sigDelay) override;
 
-        virtual void estimateSignalDelayCriticalInput(SignalDelay &sigDelay, unsigned outputPort, unsigned outputBit, unsigned &inputPort, unsigned &inputBit) override;
+        virtual void estimateSignalDelayCriticalInput(SignalDelay &sigDelay, size_t outputPort, size_t outputBit, size_t &inputPort, size_t &inputBit) override;
     protected:
         size_t m_conditionId = 0;
     };
